+++ document_crawling/document_data/postgres_module/embedding.py
